Set up for editing an idea
When a user clicks on the edit button, the current title and body are hidden,
an AJAX request is sent off for the full body text, and the static text fields
are replaced with input boxes for the user to edit. Additionally, the edit button
is hidden and is replaced with a 'Save Changes' button. There is no handler for
this button yet.